blogBackend/requirements.txt
asgiref==3.3.1
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.2
mritopng==2.2
numpy==1.20.3
Pillow==8.2.0
psycopg2-binary==2.8.6
pydicom==2.1.2
pypng==0.0.20
pytz==2021.1
sqlparse==0.4.1
